chore: create MessageTextArea and TextField components
