--- src/components/contact/ContactForm.tsx
@@ -1,83 +1,0 @@
-import React, {
-  ChangeEvent,
-  FormEvent,
-  useState,
-} from 'react';
-
-export default function ContactForm() {
-  const [formData, setFormData] = useState<{
-    email: string;
-    fullName: string;
-    message: string;
-  }>({
-    email: '',
-    fullName: '',
-    message: '',
-  });
-
-  const handleChange = (
-    e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-  ) => {
-    const { name, value } = e.target;
-    setFormData((prevData) => ({
-      ...prevData,
-      [name]: value,
-    }));
-  };
-
-  const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    // Add submission logic here...
-    setFormData({ email: '', fullName: '', message: '' });
-  };
-
-  return (
-    <form
-      onSubmit={handleSubmit}
-      className="space-y-6 max-w-md mx-auto bg-[#1AAE33] px-4 py-8 rounded-[48px]"
-    >
-      <div>
-        <input
-          type="email"
-          id="email"
-          name="email"
-          value={formData.email}
-          onChange={handleChange}
-          required
-          className="mt-1 block w-full px-3 py-2 bg-white border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-          placeholder="Name"
-        />
-      </div>
-      <div>
-        <input
-          type="text"
-          id="fullName"
-          name="fullName"
-          value={formData.fullName}
-          onChange={handleChange}
-          required
-          className="mt-1 block w-full px-3 py-2 bg-white border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-          placeholder="Email"
-        />
-      </div>
-      <div>
-        <textarea
-          id="message"
-          name="message"
-          value={formData.message}
-          onChange={handleChange}
-          required
-          rows={6}
-          className="mt-1 block w-full px-3 py-2 bg-white border border-gray-300 shadow-sm focus:outline-none focus:ring-indigo-500 rounded-xl focus:border-indigo-500 sm:text-sm"
-          placeholder="Message"
-        />
-      </div>
-      <button
-        type="submit"
-        className="w-content flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium bg-white focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-      >
-        Submit
-      </button>
-    </form>
-  );
-}
